debt: Moved authorization bits into their own duck
