+++ src/drive/mobile/ducks/authorization/actions.js
@@ -2,9 +2,9 @@ import {
   restoreCozyClientJs,
   updateBarAccessToken,
   resetClient
-} from '../lib/cozy-helper'
-import { resetPersistedState } from '../../store/persistedState'
-import { setUrl, setOffline, startReplication } from './settings'
+} from '../../lib/cozy-helper'
+import { resetPersistedState } from '../../../store/persistedState'
+import { setUrl, setOffline, startReplication } from '../../actions/settings'
 
 export const SET_TOKEN = 'SET_TOKEN'
 export const SET_CLIENT = 'SET_CLIENT'
