--- src/drive/mobile/reducers/index.js
@@ -1,11 +1,0 @@
-import { combineReducers } from 'redux'
-
-import { settings } from './settings'
-import { default as mediaBackup } from '../ducks/mediaBackup'
-import { authorization } from './authorization'
-
-export default combineReducers({
-  authorization,
-  settings,
-  mediaBackup
-})
