+++ src/drive/store/rootReducer.js
@@ -1,8 +1,16 @@
 /* global __TARGET__ */
 import { combineReducers } from 'redux'
 import baseReducers from '../reducers'
-import mobileReducer from '../mobile/reducers'
-import { UNLINK } from '../mobile/actions/authorization'
+
+import { settings } from '../mobile/reducers/settings'
+import { default as mediaBackup } from '../mobile/ducks/mediaBackup'
+import { default as authorization, UNLINK } from '../mobile/ducks/authorization'
+
+const mobileReducer = combineReducers({
+  authorization,
+  settings,
+  mediaBackup
+})
 
 // Per Dan Abramov: https://stackoverflow.com/questions/35622588/how-to-reset-the-state-of-a-redux-store/35641992#35641992
 const createRootReducer = client => {
